Dependency modifications required to run cargo test on reth-primitives
We build a lot more dependencies for a test build compared to a normal
library build, a lot of which needed to be forked for valida. We also
need to register a getrandom implementation, just like we would as if we
were creating a binary crate build.

Included for convenience is also a cargo config.toml to allow us to
simply run cargo with no extra RUSTFLAGS etc.